chore: move buildTree function to utils
